+++ src/pages/apartments/edit-apartment/apartment-policy.tsx
@@ -1,8 +1,8 @@
 import BuildingIcon from "../../../assets/icons/building";
 import { useCallback, useLayoutEffect, useMemo, useState } from "react";
-import { useAppDispatch } from "../../../stores/hooks";
+import { useAppDispatch, useAppSelector } from "../../../stores/hooks";
 import { updatePageProperties } from "../../../stores/appFunctionality/pageProperties";
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import Timeline from "../../../components/timeline";
 import AddEditApartmentPolicies from "../../../components/add-edit-apartment/apartment-policies";
 import { openSnackbar } from "../../../stores/appFunctionality/snackbar";
@@ -12,6 +12,7 @@ import useAxiosMultipart from "../../../useHooks/useAxiosMultipart";
 
 export default function EditApartmentPolicies() {
   const { id } = useParams();
+  const navigate = useNavigate();
   const axios = useAxiosMultipart();
   const breadCrumb = useMemo(
     () => [
@@ -48,25 +49,47 @@ export default function EditApartmentPolicies() {
       })
     );
   }, [breadCrumb]);
-
+  const removedImageIdSet = useAppSelector(
+    (state) => state.addEditApartmentInfo.value.data?.removeImages
+  );
   const [isSubmitting, setIsSubmitting] = useState(false);
-  const handleSubmit = useCallback(async (payload: requestPayload) => {
-    try {
-      setIsSubmitting(true);
-      const response = await axios.put(`/admin/shortlet/${id}`, payload);
-      const { shortlet } = response?.data?.data;
-      dispatch(
-        openSnackbar({
-          message: "Apartment informaton successfully updated",
-          isError: false,
-        })
+  const handleSubmit = useCallback(
+    async (payload: requestPayload) => {
+      const populatedPayload = {
+        ...payload,
+        images: payload?.images?.filter((item) => !item?.id) || [],
+        remove_images: removedImageIdSet,
+      } as {
+        [key: string]: any;
+      };
+      const newPayload = Object.fromEntries(
+        Object.entries(populatedPayload).filter(([key]) =>
+          populatedPayload[key] === "" ||
+          populatedPayload[key] === 0 ||
+          populatedPayload[key]?.length === 0
+            ? false
+            : true
+        )
       );
-      dispatch(replaceApartmentInList(shortlet));
-    } catch (error) {
-    } finally {
-      setIsSubmitting(false);
-    }
-  }, []);
+      try {
+        setIsSubmitting(true);
+        const response = await axios.post(`/admin/shortlet/${id}`, newPayload);
+        const { shortlet } = response?.data?.data;
+        dispatch(
+          openSnackbar({
+            message: "Apartment informaton successfully updated",
+            isError: false,
+          })
+        );
+        navigate(`/apartments/apartment-details/${id}`);
+        dispatch(replaceApartmentInList(shortlet));
+      } catch (error) {
+      } finally {
+        setIsSubmitting(false);
+      }
+    },
+    [removedImageIdSet, id]
+  );
 
   return (
     <section className="w-full flex flex-col items-center">